Divide into steps (debug version)
