+++ jdtls.ext/com.microsoft.jdtls.ext.core/src/com/microsoft/jdtls/ext/core/CommandHandler.java
@@ -33,7 +33,7 @@ public class CommandHandler implements IDelegateCommandHandler {
                     return PackageCommand.resolvePath(arguments, monitor);
                 case "java.project.getMainMethod":
                     return ProjectCommand.getMainMethod(arguments, monitor);
-                case "java.project.exportJar":
+                case "java.project.generateJar":
                     return ProjectCommand.exportJar(arguments, monitor);
                 default:
                     break;
